Bug Fixes (sum vs np.sum)
Modifications have been made to address the np.sum() bug. With this updated version, the optimal parameters have changed (tested with a total of 6 guests: CO2, H2O, N2O4, O2-PE, N2, C2H2).

Recommendations:
- Sigma = 0.4
- Overlap tolerance = 0.35 (primarily for larger molecules such as N2O4)

The previous version of the code smoothed the data along the x-axis only. This was achieved by scaling the data based on the ratio of the sum along the x-axis before and after applying the Gaussian filter. I have confirmed its effect on the smoothed plot and was able to replicate the data by applying the Gaussian filter solely to the x-axis.

The updated version utilizes np.sum to normalize the data across all dimensions after the Gaussian filter. As a result, it was necessary to lower the sigma value to identify reasonable binding sites accurately. For most gases, this change did not significantly affect the binding sites. However, for CO2, the lower occupancy sites were eliminated.