Occupancy Bug Fix
Update to fix the ordering issue when binding sites are printed to the XYZ file. We have added sorting to the restructured data, which should now ensure that the restructured_molecules and binding_sites variables maintain the same order.

In the process, I modified some lists. The brackets [] and occupancy were inverted, so when the coordinates were being corrected to factor in the pbc, it was inadvertently overwriting the occupancies of the restructured_molecules variable.

Additionally, I implemented the following:
      del restructured_molecules[index]
This seems correct, especially if a site is removed due to not meeting the RMSD tolerance. In such cases, the corresponding molecule should also be removed from restructured_molecules.